backend/requirements.txt
# Web 框架
fastapi>=0.112.1
uvicorn>=0.30.6

# Supabase（数据库访问 + 认证）
supabase>=2.7.2
gotrue>=2.0.0

# LLM
langchain>=0.2.14
langchain-openai>=0.1.22
langchain-experimental>=0.0.47
tiktoken>=0.5.0
dashscope>=1.25.5

# 数据验证
pydantic>=2.8.2

# 工具
python-dotenv>=1.0.1
httpx>=0.27.0

# RSS 解析
feedparser>=6.0.0

# 任务队列
celery[redis]>=5.3.0
redis>=5.0.0
flower>=2.0.0

email-validator>=2.0.0

# Encryption
cryptography>=41.0.0

# Image processing
Pillow>=10.0.0
beautifulsoup4>=4.12.0